refactor(suite-native): rename dynamic header files
